+++ changelog.md
@@ -3,6 +3,7 @@
 ## [1.11]
 - new calendrical type synonyms and abstract constructors
 - new functions for working with week-based years
+- new parseTimeMultipleM function for a list of (format, input) pairs
 - add instance Ord DayOfWeek
 - add instance Read DiffTime (and NominalDiffTime)
 - change instance Read UTCTime to allow omitted timezone